src/main/java/com/example/shoppingwebapplication/Services/ProductServiceImpl.java
package com.example.shoppingwebapplication.Services;

import com.example.shoppingwebapplication.Entity.Product;
import com.example.shoppingwebapplication.Repository.ProductRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.stereotype.Service;

@Service
public class ProductServiceImpl implements ProductService {
    private final ProductRepository productRepository;

    @Autowired
    public ProductServiceImpl(ProductRepository productRepository) {
        this.productRepository = productRepository;
    }

    @Override
    public Page<Product> getAllProducts(Pageable page) {
        return productRepository.findAll(page);
    }

    @Override
    public void saveProduct(Product product) {
        productRepository.save(product);
    }

    @Override
    public Product findProductById(int productId) {
        return productRepository.findById(productId);
    }
}
